requirements.txt
numpy>=1.21.0
Pillow>=8.0.0
requests>=2.25.0
scikit-learn>=1.0.0
